fix: observe Task column measurement changes
