+++ apps/desktop/src/features/home/ProjectTaskLabelsCell.tsx
@@ -9,18 +9,8 @@ import {
   TaskLabelAssignmentProvider,
   useTaskLabelAssignment,
 } from "@/shared/labels";
-import { Badge, OneLineOverflowRow, Spinner, type OneLineOverflowItem } from "@/ui";
-
-export function projectTaskLabelItems(labels: TaskListItem["labels"]): readonly OneLineOverflowItem[] {
-  return labels.map((label) => ({
-    content: (
-      <Badge className="py-[3px]" size="compact" tone="neutral">
-        {label.name}
-      </Badge>
-    ),
-    id: label.id,
-  }));
-}
+import { Badge, OneLineOverflowRow, Spinner } from "@/ui";
+import { projectTaskLabelItems } from "./projectTaskLabelItems";
 
 export function ProjectTaskLabelsCell({
   onOpenChange,
